# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JustEnough API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/inventory/snapshots": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Get Snapshots",
        "operationId": "get_snapshots_api_v1_inventory_snapshots_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "branch_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Branch Id"
            }
          },
          {
            "name": "product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Create Snapshot",
        "operationId": "create_snapshot_api_v1_inventory_snapshots_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/movements": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Get Movements",
        "operationId": "get_movements_api_v1_inventory_movements_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/recommendations": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Recommendations",
        "operationId": "get_recommendations_api_v1_recommendations_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "branch_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Branch Id"
            }
          },
          {
            "name": "product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Create Recommendation",
        "operationId": "create_recommendation_api_v1_recommendations_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{id}/override": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Override Recommendation",
        "operationId": "override_recommendation_api_v1_recommendations__id__override_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{id}/explain": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Explain Recommendation",
        "operationId": "explain_recommendation_api_v1_recommendations__id__explain_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OverrideRequest": {
        "properties": {
          "override_reason": {
            "type": "string",
            "minLength": 3,
            "title": "Override Reason"
          },
          "recommended_qty": {
            "type": "number",
            "minimum": 0.0,
            "title": "Recommended Qty"
          }
        },
        "type": "object",
        "required": [
          "override_reason",
          "recommended_qty"
        ],
        "title": "OverrideRequest"
      },
      "RecommendationCreateRequest": {
        "properties": {
          "branch_id": {
            "type": "string",
            "title": "Branch Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "target_date": {
            "type": "string",
            "format": "date",
            "title": "Target Date"
          },
          "recommended_qty": {
            "type": "number",
            "minimum": 0.0,
            "title": "Recommended Qty"
          },
          "risk": {
            "type": "string",
            "title": "Risk",
            "default": "MEDIUM"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "branch_id",
          "product_id",
          "target_date",
          "recommended_qty"
        ],
        "title": "RecommendationCreateRequest"
      },
...
      "SnapshotCreateRequest": {
        "properties": {
          "branch_id": {
            "type": "string",
            "title": "Branch Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "snapshot_date": {
            "type": "string",
            "format": "date",
            "title": "Snapshot Date"
          },
          "quantity": {
            "type": "number",
            "minimum": 0.0,
            "title": "Quantity"
          },
          "data_flag": {
            "type": "string",
            "title": "Data Flag",
            "default": "ACTUAL"
          }
        },
        "type": "object",
        "required": [
          "branch_id",
          "product_id",
          "snapshot_date",
          "quantity"
        ],
        "title": "SnapshotCreateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        }
      }
    }
  }
}